# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Db server API description.",
        "title": "Db server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://openitstudio.ru",
            "email": "fedor@support-pc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/admin/pl": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List pipelines",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline",
                    "Admin"
                ],
                "summary": "List pipelines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pipeline.Pipeline"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create pipeline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline",
                    "Admin"
                ],
                "summary": "Create pipeline",
                "parameters": [
                    {
                        "description": "Pipeline info",
                        "name": "pl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.Pipeline"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pipeline.Pipeline"
                        }
                    }
                }
            }
        },
        "/admin/pl/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Pipeline detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline",
                    "Admin"
                ],
                "summary": "Pipeline info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pipeline.Pipeline"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update pipeline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline",
                    "Admin"
                ],
                "summary": "Update pipeline",
                "parameters": [
                    {
                        "description": "Pipeline info",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.Pipeline"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Pipeline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pipeline.Pipeline"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete pipeline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline",
                    "Admin"
                ],
                "summary": "Delete pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/user/oauth/{provider}/link": {
            "get": {
                "description": "Get link for oauth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth",
                    "Public Api"
                ],
                "summary": "OAuth link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth key",
                        "name": "db-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/oauth/{provider}/{code}": {
            "get": {
                "description": "Get user by oauth code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth",
                    "Public Api"
                ],
                "summary": "OAuth user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth key",
                        "name": "db-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pipeline.Pipeline": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The pipeline UUID\nexample: 6204037c-30e6-418b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "input": {
                    "description": "Input type",
                    "type": "string"
                },
                "input_id": {
                    "description": "The pipeline input UUID\nexample: 6204037c-30e6-418b-8saa-dd8219860b4b",
                    "type": "string"
                },
                "output": {
                    "description": "Output type",
                    "type": "string"
                },
                "output_id": {
                    "description": "The pipeline output UUID\nexample: 6204037c-30e6-413b-8saa-dd8219860b4b",
                    "type": "string"
                },
                "title": {
                    "description": "Mnemonic name",
                    "type": "string"
                }
            }
        },
...
        "user.User": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Is user active",
                    "type": "boolean"
                },
                "admin": {
                    "description": "Is user admin",
                    "type": "boolean"
                },
                "created_at": {
                    "description": "Created at date time",
                    "type": "string"
                },
                "devices": {
                    "description": "user devices",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserDevice"
                    }
                },
                "email": {
                    "description": "User email",
                    "type": "string"
                },
                "id": {
                    "description": "The user UUID\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "last_login": {
                    "description": "Last login date time",
                    "type": "string"
                },
                "token": {
                    "description": "Auth token",
                    "type": "string"
                }
            }
        },
        "user.UserDevice": {
            "type": "object",
            "properties": {
                "device": {
                    "description": "Device type can be android | ios | macos | windows | linux | web\nexample: 'linux'",
                    "type": "string"
                },
                "device_token": {
                    "description": "UUID Device Token\n9f80fcc8-0102-4795-94c0-4190c168ffc2",
                    "type": "string"
                },
                "id": {
                    "description": "The UUID of a device\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "user_id": {
                    "description": "The UUID of owned user\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}